Refactor file names for clarity during bonus mode
With this commit, file names for bonus mode have been refactored for clarity. A couple of file name changes and redefinitions have been carried out in the gamepad and minimap sections. Changes include renaming 'minimap.h' to 'minimap_bonus.h', 'gamepad.h' to 'gamepad_bonus.h', and updating all references in other files accordingly. The 'create_map' function is moved from 'minimap.h' to 'minimap_bonus.h', while the 'minimap_draw_player' function is deleted.

This was done to make the distinctions between regular and bonus mode clearer, thus facilitating better code readability and maintainability. This change will not affect the functionality of the game."
