+++ bonus/srcs/gamepad/input_loop_bonus.c
@@ -1,27 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   clear_gamepads.c                                   :+:      :+:    :+:   */
+/*   input_loop.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: beroux <beroux@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/25 05:31:29 by beroux            #+#    #+#             */
-/*   Updated: 2023/09/25 16:03:30 by beroux           ###   ########.fr       */
+/*   Created: 2023/09/20 04:08:09 by beroux            #+#    #+#             */
+/*   Updated: 2023/09/25 07:26:54 by beroux           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "gamepad.h"
+#include "gamepad_bonus.h"
 
-void	clear_gamepads(t_gamepad **list)
+int	input_loop(t_gamepad *gamepad)
 {
-	if (!list || !*list)
-		return ;
-	if ((*list)->next)
-		clear_gamepads(&(*list)->next);
-	if ((*list)->name)
-		free((*list)->name);
-	if ((*list)->fd != -1)
-		close((*list)->fd);
-	free(*list);
-	*list = NULL;
+	t_gamepad	*tmp;
+
+	while (1)
+	{
+		update_gamepads(gamepad);
+		tmp = gamepad;
+		while (tmp)
+		{
+			update_inputs(tmp);
+			tmp = tmp->next;
+		}
+		usleep(10000);
+	}
+	return (0);
 }
