+++ bonus/srcs/gamepad/sources.mk
@@ -1,9 +1,9 @@
-SRCS +=	gamepad/hooks/button_hooks.c \
-		gamepad/hooks/axis_hook.c \
-		gamepad/update_gamepads.c \
-		gamepad/init_gamepads.c \
-		gamepad/clear_gamepads.c \
-		gamepad/update_inputs.c \
-		gamepad/input_loop.c \
+SRCS +=	gamepad/hooks/button_hooks_bonus.c \
+		gamepad/hooks/axis_hook_bonus.c \
+		gamepad/update_gamepads_bonus.c \
+		gamepad/init_gamepads_bonus.c \
+		gamepad/clear_gamepads_bonus.c \
+		gamepad/update_inputs_bonus.c \
+		gamepad/input_loop_bonus.c \
 
-INCS += incs/gamepad.h
+INCS += incs/gamepad_bonus.h
