+++ bonus/srcs/player/input_player_js_bonus.c
