+++ bonus/srcs/player/sources.mk
@@ -1,6 +1,6 @@
 SRCS += player/init_player_bonus.c \
 		player/update_player_bonus.c \
 		player/input_player_bonus.c \
-		player/input_player_js.c \
+		player/input_player_js_bonus.c \
 
 INCS += incs/player_bonus.h
